feat(asserters) Extract asserters from constraints.
Instead of declaring specific asserters inside the constraint files,
they are now extracted into their own files in the
`atoum\phpunit\asserters` namespace. The naming is different too to
mimic atoum's asserters structure.
